Rename AppNotificationChannels to NotificationChannels, and move to NotificationUtils.kt
